Clean up videoRenderer
