--- StreamVideoSwiftUITests/Utils/PiPUtils_Tests.swift
@@ -1,77 +1,0 @@
-//
-// Copyright © 2024 Stream.io Inc. All rights reserved.
-//
-
-@testable import StreamVideo
-@testable import StreamVideoSwiftUI
-import StreamSwiftTestHelpers
-import XCTest
-import CoreMedia
-import SnapshotTesting
-
-final class PiPUtils_Tests: XCTestCase {
-    
-    func test_sampleBuffer_fromPixelBuffer() {
-        // Given
-        let image = UIImage(named: "test", in: Bundle(for: type(of: self)), with: nil)!
-        let buffer = buffer(from: image)!
-        
-        // When
-        let sampleBuffer = CMSampleBuffer.from(buffer)!
-        let imageBuffer = CMSampleBufferGetImageBuffer(sampleBuffer)!
-        let ciimage = CIImage(cvPixelBuffer: imageBuffer)
-        let converted = self.convert(cmage: ciimage)
-        
-        // Then
-        assertSnapshot(matching: converted, as: .image(precision: precision))
-    }
-    
-    func buffer(from image: UIImage) -> CVPixelBuffer? {
-        let attrs = [
-            kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue,
-            kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue
-        ] as CFDictionary
-        
-        var pixelBuffer : CVPixelBuffer?
-        let status = CVPixelBufferCreate(
-            kCFAllocatorDefault, Int(image.size.width),
-            Int(image.size.height),
-            kCVPixelFormatType_32ARGB,
-            attrs, &pixelBuffer
-        )
-        guard (status == kCVReturnSuccess) else {
-            return nil
-        }
-        
-        CVPixelBufferLockBaseAddress(pixelBuffer!, CVPixelBufferLockFlags(rawValue: 0))
-        let pixelData = CVPixelBufferGetBaseAddress(pixelBuffer!)
-        
-        let rgbColorSpace = CGColorSpaceCreateDeviceRGB()
-        let context = CGContext(
-            data: pixelData,
-            width: Int(image.size.width),
-            height: Int(image.size.height),
-            bitsPerComponent: 8,
-            bytesPerRow: CVPixelBufferGetBytesPerRow(pixelBuffer!),
-            space: rgbColorSpace,
-            bitmapInfo: CGImageAlphaInfo.noneSkipFirst.rawValue
-        )
-        
-        context?.translateBy(x: 0, y: image.size.height)
-        context?.scaleBy(x: 1.0, y: -1.0)
-        
-        UIGraphicsPushContext(context!)
-        image.draw(in: CGRect(x: 0, y: 0, width: image.size.width, height: image.size.height))
-        UIGraphicsPopContext()
-        CVPixelBufferUnlockBaseAddress(pixelBuffer!, CVPixelBufferLockFlags(rawValue: 0))
-        
-        return pixelBuffer
-    }
-    
-    func convert(cmage: CIImage) -> UIImage {
-        let context = CIContext(options: nil)
-        let cgImage = context.createCGImage(cmage, from: cmage.extent)!
-        let image = UIImage(cgImage: cgImage)
-        return image
-    }
-}
